klms-cancel-fetcher/requirements.txt
python-dotenv
requests
openai
fastapi
uvicorn 